No need to seperate models again
